+++ nix/docs/docs-man/package.nix
@@ -13,6 +13,8 @@ let
   helpers = {
     mkSkip = ../../build-support/mkSkip.nix;
     mkEntries = ../../build-support/mkEntries.nix;
+    mkEvalChecks = ../../build-support/mkEvalChecks.nix;
+    mkReport = ../../build-support/mkReport.nix;
     mkResultWith = ../../build-support/mkResultWith.nix;
     mkResult = ../../build-support/mkResult.nix;
     mkSnapshot = ../../build-support/mkSnapshot.nix;